Breadth-First Search (BFS) Algorithm With Python
BFS goes wide. That means we traverse a graph level by level. We take a node, traverse all of its child nodes, and then move to another node. Because all of the child nodes of a node are on the same level.

Now let's understand BFS more thoroughly by visualizing each step of the algorithm using this graph:

Let’s see how the level-wise traversing works in this graph considering ‘A’ as the root node:

Queue for Level-Wise Search
Queue is a linear structure of data that follows the First-In-First-Out (FIFO) principle. The element entering first in a queue will also leave first.

The FIFO principle of a queue will ensure the level-wise searching of BFS. Let’s see how from the sketch below: